bump nanobind version: we now have support for fp8 dtypes
